requirements.txt
absl-py==0.7.1
astor==0.7.1
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
certifi==2019.6.16
cycler==0.10.0
daal4py==0.2019.1.1
decorator==4.3.0
emcee==3.0rc2
gast==0.2.2
google-pasta==0.1.7
grpcio==1.16.1
h5py==2.8.0
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.12.0
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
Markdown==3.1.1
matplotlib==2.2.2
MCOrbit==0.0.0
mkl-fft==1.0.2
mkl-random==1.0.1
mock==3.0.5
numexpr==2.6.5
numpy==1.15.4
olefile==0.46
pandas==0.23.0
parso==0.2.0
path.py==11.0.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==6.0.0
prompt-toolkit==1.0.15
protobuf==3.8.0
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2018.4
PyYAML==4.1
scikit-learn==0.21.1
scipy==1.1.0
SciServer==2.0.13
simplegeneric==0.8.1
six==1.11.0
tables==3.4.3
TBB==0.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.11.1
